requirements.txt
Django==2.1.2
future==0.17.1
iso8601==0.1.12
mysql==0.0.2
mysqlclient==1.3.13
pkg-resources==0.0.0
pyserial==3.4
pytz==2018.6
PyYAML==3.13
vine==1.2.0
